requirements.txt
Flask
hypercorn
paho-mqtt
websockets
Pillow